requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-huggingface>=0.0.8
langchain-chroma>=0.1.2
langchain-community>=0.3.0
chromadb>=0.5.0
pandas>=2.2.0
openpyxl>=3.1.2
streamlit>=1.38.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
pytest>=8.3.0
google-generativeai>=0.8.0
google-genai>=1.0.0
tf-keras
